Move filters package to the API.
These filters are only use to interchange data between clients and daemons.
They don't belong to the parsers package.
