+++ api/client/client.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/docker/docker/api/client/lib"
 	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/api/types/registry"
-	"github.com/docker/docker/pkg/parsers/filters"
 	"github.com/docker/docker/runconfig"
 )
