+++ api/client/lib/container_list.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/pkg/parsers/filters"
+	"github.com/docker/docker/api/types/filters"
 )
 
 // ContainerList returns the list of containers in the docker host.
