+++ api/server/router/system/backend.go
@@ -2,8 +2,8 @@ package system
 
 import (
 	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/pkg/jsonmessage"
-	"github.com/docker/docker/pkg/parsers/filters"
 )
 
 // Backend is the methods that need to be implemented to provide
