+++ integration-cli/docker_api_network_test.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 
 	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/filters"
 	"github.com/docker/docker/api/types/network"
 	"github.com/docker/docker/pkg/integration/checker"
-	"github.com/docker/docker/pkg/parsers/filters"
 	"github.com/go-check/check"
 )
